+++ typescript/projects/Rylogic.TextAligner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.11.1] - 2022-03-05
+
+- Added 'Ignore Line Pattern' option
+- Bug Fix. The user setting for 'alignCharacters' was not being read correctly
+- Synchronised with VS release
+
 ## [1.10.6] - 2021-09-13
 
 - Rebuild with updated packages. No other changes.